requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo
pymongo[srv]==3.12
scikit-learn
pyaml
dill
mlflow

##-e .